sonarqube fixes

+++ iabgpp-encoder/src/main/java/com/iab/gpp/encoder/datatype/encoder/BaseIntegerSet.java
@@ -18,7 +18,7 @@ public abstract class BaseIntegerSet extends AbstractSet<Integer> implements Int
     if (value == null) {
       return false;
     }
-    return addInt((Integer) value);
+    return addInt(value);
   }
   
   @Override
